Move config to logging mod
